POH-915 - WIP - flyttet BeløpBeregning fra utils
